finished auth

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,21 @@
 import { UseWorkoutsContext } from "../hooks/UseWorkoutsContext"
+import { useAuthContext } from "../hooks/UseAuthContext" 
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = UseWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
         if (response.ok) {
